requirements.txt
numpy
opencv-python
mediapipe
Pillow
trimesh
matplotlib
scipy
tkinterweb
